# web/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Referral API",
        "description": "This is API Microservice Referral Program",
        "contact": {
            "name": "Support Team",
            "email": "admin@sumra.net"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/referrals": {
            "get": {
                "tags": [
                    "Main"
                ],
                "summary": "List all referrals for user",
                "description": "List all referrals for user",
                "operationId": "App\\Api\\V1\\Controllers\\MainController::index",
                "responses": {
                    "200": {
                        "description": "Success join new user to referrer user"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "default": [
                            "ManagerRead",
                            "User",
                            "ManagerWrite"
                        ]
                    }
                ],
                "x-auth-type": "Application & Application User",
                "x-throttling-tier": "Unlimited",
                "x-wso2-application-security": {
                    "security-types": [
                        "oauth2"
                    ],
                    "optional": "false"
                }
            },
            "post": {
                "tags": [
                    "Main"
                ],
                "summary": "Join new user to referrer",
                "description": "Send encryption data: username - New User Name, * package_name (required) - Package Name, referrer_code - Referrer code, * device_id (required) - User Device ID, * device_name (required) - User Device Name",
                "operationId": "App\\Api\\V1\\Controllers\\MainController::create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "version_key": {
                                        "description": "Version Key",
                                        "type": "string",
                                        "example": "66981685"
                                    },
                                    "data": {
                                        "description": "Encrypt data",
                                        "type": "text",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success send data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "default": [
                            "ManagerRead",
                            "User",
                            "ManagerWrite"
                        ]
                    }
                ],
                "x-auth-type": "Application & Application User",
                "x-throttling-tier": "Unlimited",
                "x-wso2-application-security": {
                    "security-types": [
                        "oauth2"
                    ],
                    "optional": "false"
                }
            }
        },
        "/v1/referrals/invite": {
            "get": {
                "tags": [
                    "Main"
                ],
                "summary": "Get user invite code",
                "description": "Get user referrer invite code",
                "operationId": "App\\Api\\V1\\Controllers\\MainController::invite",
                "parameters": [
                    {
                        "name": "package_name",
                        "in": "query",
                        "description": "Package Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "net.sumra.chat"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success get or generate referrer invite code"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "default": [
                            "ManagerRead",
                            "User",
                            "ManagerWrite"
                        ]
                    }
                ],
                "x-auth-type": "Application & Application User",
                "x-throttling-tier": "Unlimited",
                "x-wso2-application-security": {
                    "security-types": [
                        "oauth2"
                    ],
                    "optional": "false"
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "default": {
                "type": "oauth2",
                "description": "Auth Scheme",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "https://is.sumra.net/oauth2",
                        "scopes": {
                            "ManagerRead": "Manager can read",
                            "User": "User access",
                            "ManagerWrite": "Manager can write"
                        }
                    }
                }
            }
        }
    }
}